backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pydantic[email]==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0

# Rate limiting
slowapi==0.1.9

# Auth / security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
pyotp==2.9.0

# Email
aiosmtplib==3.0.2
jinja2==3.1.6

# Object storage (S3-compatible)
boto3==1.38.17

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
